Separate app definition and controllers in examples
Start off doing this with the examples to get an idea of what this would look like. Have to rethink how the `route` decorator works now that I know it's not possible to use the `app` directly through `definition` for initializing models/running the server because the controllers are defined in the other file after importing from the definition. Or maybe rethink what this separation looks like.